Comments, Formatting
A couple more comments have been added, and default comments have been
removed for now. A little bit of formatting has been changed for
comments and a couple variables and functions. The useless key code has
been reduced to one int variable and is now more reliable, faster, and
lighter on the bot as to prevent any slowing down of the code. This may
be the final update until Bag Day.